@@ -1,57 +1,57 @@
 import * as APIUtil from '../util/session_api_util';
-import jwt_decode from 'jwt-decode';
+import { jwtDecode } from 'jwt-decode';
 
 export const RECEIVE_CURRENT_USER = "RECEIVE_CURRENT_USER";
 export const RECEIVE_SESSION_ERRORS = "RECEIVE_SESSION_ERRORS";
 export const RECEIVE_USER_LOGOUT = "RECEIVE_USER_LOGOUT";
 
 export const receiveCurrentUser = currentUser => ({
-    type: RECEIVE_CURRENT_USER,
-    currentUser
+  type: RECEIVE_CURRENT_USER,
+  currentUser
 });
 
 export const receiveErrors = errors => ({
-    type: RECEIVE_SESSION_ERRORS,
-    errors
+  type: RECEIVE_SESSION_ERRORS,
+  errors
 });
 
 export const logoutUser = () => ({
-    type: RECEIVE_USER_LOGOUT
+  type: RECEIVE_USER_LOGOUT
 });
 
 export const signup = user => dispatch => (
-    APIUtil.signup(user).then(res => {
-        const { token } = res.data;
-        localStorage.setItem('jwtToken', token);
-        APIUtil.setAuthToken(token);
-        const decoded = jwt_decode(token);
-        dispatch(receiveCurrentUser(decoded))
-    }) 
-        .catch(err => (
-            dispatch(receiveErrors(err.response.data))
-        ))
+  APIUtil.signup(user).then(res => {
+    const { token } = res.data;
+    localStorage.setItem('jwtToken', token);
+    APIUtil.setAuthToken(token);
+    const decoded = jwtDecode(token);
+    dispatch(receiveCurrentUser(decoded))
+  }) 
+    .catch(err => (
+        dispatch(receiveErrors(err.response.data))
+    ))
 );
 
 export const login = user => dispatch => (
-    APIUtil.login(user).then(res => {
-        const { token } = res.data;
-        localStorage.setItem('jwtToken', token);
-        APIUtil.setAuthToken(token);
-        const decoded = jwt_decode(token);
-        dispatch(receiveCurrentUser(decoded))
+  APIUtil.login(user).then(res => {
+    const { token } = res.data;
+    localStorage.setItem('jwtToken', token);
+    APIUtil.setAuthToken(token);
+    const decoded = jwtDecode(token);
+    dispatch(receiveCurrentUser(decoded))
+  })
+    .catch(err => {
+      dispatch(receiveErrors(err.response.data));
     })
-        .catch(err => {
-            dispatch(receiveErrors(err.response.data));
-        })
 )
 
-export const update = user => dispatch => (
-    APIUtil.update(user)
-    .catch(err => console.log(err))
+export const update = user => () => (
+  APIUtil.update(user)
+         .catch(err => console.log(err))
 )
 
 export const logout = () => dispatch => {
-    localStorage.removeItem('jwtToken')
-    APIUtil.setAuthToken(false)
-    dispatch(logoutUser())
+  localStorage.removeItem('jwtToken')
+  APIUtil.setAuthToken(false)
+  dispatch(logoutUser())
 };
