@@ -13,7 +13,7 @@ function WordRulesModal () {
         </li>
           <ul className="ways-list">
             <li><span className='bullet'>A.</span> Click on each tile individually, in order. When the word is fully spelled, click the green ✔ in the Word Bank.</li>
-            <li><span className='bullet'>B.</span> Press the mouse button down over a word's first letter and drag the cursor over the subsequent letters. Release the mouse button to submit.</li>
+            <li><span className='bullet'>B.</span> Press the mouse button down over a word&apos;s first letter and drag the cursor over the subsequent letters. Release the mouse button to submit.</li>
           </ul>
         <li>
           <span className='bullet'>2.</span> You cannot select a letter that has already been selected for the current word.
