@@ -6,15 +6,15 @@ import {
 const _nullErrors = [];
 
 const SessionErrorsReducer = (state = _nullErrors, action) => {
-    Object.freeze(state);
-    switch (action.type) {
-        case RECEIVE_SESSION_ERRORS:
-            return action.errors;
-        case RECEIVE_CURRENT_USER:
-            return _nullErrors;
-        default:
-            return state;
-    }
+  Object.freeze(state);
+  switch (action.type) {
+    case RECEIVE_SESSION_ERRORS:
+      return action.errors;
+    case RECEIVE_CURRENT_USER:
+      return _nullErrors;
+    default:
+      return state;
+  }
 };
 
 export default SessionErrorsReducer;
