@@ -1,16 +1,19 @@
-import { createStore, applyMiddleware } from 'redux';
+import { createStore, applyMiddleware, compose } from 'redux';
 import thunk from 'redux-thunk';
-import rootReducer from '../reducers/rootReducer';
+import rootReducer from "../reducers/rootReducer";
 
-const middlewares = [thunk];
-
-if (process.env.NODE_ENV !== 'production') {
-    const { logger } = require('redux-logger');
-    middlewares.push(logger);
+let enhancer;
+if (import.meta.env.MODE === "production") {
+  enhancer = applyMiddleware(thunk);
+} else {
+  const logger = (await import("redux-logger")).default;
+  const composeEnhancers =
+    window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
+  enhancer = composeEnhancers(applyMiddleware(thunk, logger));
 }
 
-const configureStore = (preloadedState = {}) => {
-    return createStore(rootReducer, preloadedState, applyMiddleware(...middlewares))
+const configureStore = (preloadedState) => {
+  return createStore(rootReducer, preloadedState, enhancer);
 };
 
 export default configureStore;
